pyproject.toml
[project]
name = "self-evolving-game-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "open_spiel",
    "pyyaml",
    "openai",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-timeout"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
